# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Parmana API
  version: 1.0.0
  description: >
    Parmana is an Execution Trust Infrastructure that ensures there is no gap
    between what humans decide and what AI systems do. The API enables creation,
    execution, verification, replay, and auditing of Business Transactions
    through cryptographically verifiable Execution Trust Records.


    **Every route requires a caller bearer key, except GET /health.** Send
    `Authorization: Bearer <key>` on every other request. Keys are issued by
    `scripts/generate-api-key.ts` and configured server-side via
    `PARMANA_API_KEYS`; only a hash of each key is ever held by the server,
    verified in constant time. A missing or invalid credential returns 401
    before a Business Transaction is even constructed, independent of Policy
    evaluation and gateway attestation, see `packages/api/src/middleware/caller-auth.ts`
    and [Authentication](/api-reference/authentication). Local development may
    set `PARMANA_AUTH_DISABLED=true` to skip this middleware entirely; that flag
    must never be set in a real deployment.

  contact:
    name: Parmana
    email: support@parmana.ai

  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:3000
    description: Local (packages/api, PORT env var, default 3000)
security:
  - bearerAuth: []
...
paths:
...
  /transactions/{businessTransactionId}:
    get:
      tags:
        - Transactions
      summary: Get Business Transaction
      operationId: getTransaction
      description: Returns a bare Business Transaction by ID. No format validation is applied to the path parameter.
      parameters:
        - name: businessTransactionId
          in: path
          required: true
          description: Unique Business Transaction identifier.
          schema:
            type: string
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Business Transaction found.
          content:
            application/json:
              schema:
                $ref: "../schemas/responses/transaction-response.schema.json"
              examples:
                vendor-payment-flow-transactions-endpoint:
                  summary: Real captured response
                  value:
                    businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                    metadata:
                      businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                      correlationId: cc1ca975-4935-4e9a-a591-8897af6b56fb
                      sourceSystem: vendor-payment-service
                      submittedBy: ap-automation
                      submittedAt: "2026-07-13T17:36:00.835Z"
                    authority:
                      authorityId: db5c5260-bb4c-40a2-a081-2928b5fe7720
                      authorityType: SERVICE
                      principalId: ap-automation-svc
                      displayName: Accounts Payable Automation
                      issuedAt: "2026-07-13T17:36:00.835Z"
                    authorization:
                      authorizationId: 63bdc429-14e4-4ef4-8007-f3c0c9564261
                      authorityId: db5c5260-bb4c-40a2-a081-2928b5fe7720
                      purpose: Authorize vendor payment disbursement
                      issuedAt: "2026-07-13T17:36:00.835Z"
                    intent:
                      intentId: 1e77d0b3-2e57-4e31-b780-03abc92f1b69
                      authorizationId: 63bdc429-14e4-4ef4-8007-f3c0c9564261
                      action: payments:execute
                      target: vendor/V-500
                      parameters:
                        amount: 5000
                        currency: USD
                      createdAt: "2026-07-13T17:36:00.835Z"
                    policy:
                      name: vendor-payment
                      version: 2.0.0
                      schemaVersion: 1.0.0
                    signals:
                      vendorVerified: true
                      invoiceVerified: true
                      paymentApproved: true
                      sufficientFunds: true
                      paymentAmount: 5000
                      riskScore: 12
                    status: RECEIVED
                    createdAt: "2026-07-13T17:36:00.965Z"

        "404":
          description: Business Transaction not found.
          content:
            application/json:
              schema:
                $ref: "../schemas/common/error.schema.json"
              examples:
                notFound:
                  summary: Real captured response
                  value:
                    error: Business Transaction not found.
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Caller API key issued by scripts/generate-api-key.ts. Sent as
        Authorization: Bearer <key>. Verified against a stored SHA-256 hash
        in constant time by packages/api/src/auth/StaticKeyAuthenticator.ts.
        Required on every route except GET /health. See
        /api-reference/authentication.
  responses:
    Unauthorized:
      description: >-
        Missing or invalid caller credential (StaticKeyAuthenticator returned
        no identity). Real captured response, packages/api/src/middleware/caller-auth.ts.
      content:
        application/json:
          schema:
            $ref: "../schemas/common/error.schema.json"
          examples:
            authRequired:
              summary: Real captured response, missing or invalid Authorization header
              value:
                error: authentication required